Add obj_get and array_get readonlyness when used as an lvalue
Summary:
During an assignment, we sometimes need to explicitly add the readonlyness of the RHS and LHS.

For a property assignment, ie.

```
$x->prop = <rhs>;
```

if RHS is readonly (and $x is mutable), the runtime needs to check that prop is a readonly property of $x.

For array assignment, i.e.:

```
$<lval>[0] = <rhs>;
```

we need to track two things: if <lval> is readonly then it must be a value type collection like vec or dict. If <lval> is mutable, <rhs> must not be readonly. Thus we need explicit readonly expressions on both sides.

Note that we do not always know the readonlyness of <lval> during HackC compile time, ie, the LHS can be an expression like
```
($x->prop->prop2)[0] = <rhs>;
```
thus we depend on the runtime to enforce this when <rhs> is readonly.

Reviewed By: kavoor

Differential Revision: D30057455

fbshipit-source-id: 79cded08431d3ca961979c5ffabed73806ecc81d