Manually fix remaining HHVM tests broken by turning on comparison coercion notices
Summary: While this may technically violate the spirit of some of the tests, the tradeoff here is that when, in a later diff in the stack, we rip out the logic that allowed for coercions, we gain much more than we lose by having a number of tests that may no longer serve as much purpose as before. Thus, we determine that the tradeoff in which the tests are slightly worsened in exchange for a much quicker cleanup and path to core logic fixes is worth it.

Differential Revision: D29947058

fbshipit-source-id: 5577068feaa9ddf9d15f0a634b6dfe954499015e