ReadOnlyOp CheckMutROCOW
Summary:
We need one more ReadOnlyOp for this case:
```
$x[0] = readonly $a;
```

Currently, `CheckROCOW` will check that "Must be mutable unless readonly + COW". This was for checking $x in the case
```
$x[0] = $a;
```
where we want to be able to modify readonly value-type arrays in the lhs. James brought up the first case, and it occurred to me that we also need to support the case we are reading a readonly value into a readonly value-type array, and for HHVM to assert this array is readonly.

Reviewed By: oulgen

Differential Revision: D30072575

fbshipit-source-id: 62a37b3ac58c319ccf0302b5dfc5ecb503ab7316