Fix slow parsing stress test
Summary: This test was supposed to be a stress test of the parser, parsing a giant string literal. Instead, it turned into a stress test of string-to-int coercion. FIXME\neq is so much slower than !== that the overall test runtime went up by 4x in dev mode, from ~75s to ~300s, triggering timeouts.

Reviewed By: ricklavoie, DavidSnider

Differential Revision: D30112942

fbshipit-source-id: 313aff341b4af5160f31af9ada9373133e1a5496