Never return 'array' from gettype
Summary:
Pre-HADVAs, gettype returned "array" for PHP arrays. As a result, we needed to preserve this behavior for legacy arrays post flag-flip.

It's ugly that gettype now returns "dict" for some dicts, and "array" for others. Fix that. Anyone depending on this behavior can shim it in user-land with an array_mark_legacy test:

```
function old_and_busted_gettype(mixed $x): string {
  return HH\is_array_marked_legacy($x) ? 'array' : \gettype($x);
}
```

Reviewed By: ricklavoie

Differential Revision: D30082038

fbshipit-source-id: 51f3faffa9991a03f9dbc921e7bc1730a7f8f270