prefer `string` or `int` to `arraykey` when adding a `Hole` on append to `keyset` or `Set`
Summary:
When appending an element to a `keyset` or `Set` we check that the value can be giving a type which is a subtype of `arraykey` adding a `Hole` with that as the expected type if it cannot.

However, when applying the unsafe cast codemod, the resulting collection will now have a value type of `arraykey` which can cause cascading errors.

If the value type is an optional `int` or `string`, the common case for this `Hole`, we can use the more specific type instead of `arraykey`.

Reviewed By: CatherineGasnier

Differential Revision: D29815086

fbshipit-source-id: c9d0f95b783cf9c4cb379316e952b4ecd28945eb