Log when we seem to be waiting indefinitely for a killed server
Summary:
There's a bug when restarting the server from the monitor.

We kill the server by sending a SIGUSR2 to the server. If the server was only recently launched, the signal handler might not yet be installed. As such the server ignores the signal, and the monitor waitpids forever on the server.

Let's log when this happens, so that we can be sure we've fixed it.

Reviewed By: ljw1004

Differential Revision: D30132610

fbshipit-source-id: d86e2295a157404639e8576ee7fb9e1777f98220