Keep killing server when monitor_kill_again_fix=true
Summary:
There's a bug when restarting the server from the monitor.

We kill the server by sending a SIGUSR2 to the server. If the server was only recently launched, the signal handler might not yet be installed. As such the server ignores the signal, and the monitor waitpids forever on the server.

In this diff, when `monitor_kill_again_fix` is enabled, it'll do:

```
forever {
   kill;
   for at most 2 seconds {
      waitpid;
   }
}
```

Reviewed By: ljw1004

Differential Revision: D30133209

fbshipit-source-id: e17a0b94e3e8697583e46f10dd4bb6d5da4331c7