Clarify names and types in 'missing member' error logic
Summary:
Previously, we used `did_you_mean` and `closest`, but it wasn't obvious which corresponded to a static versus instance method. Use explicit `static` or `instance`, and use `None` when there are no obvious things we can suggest.

Also prefer "suggestion" over "hint", as we use "hint" elsewhere in the codebase to refer to the user-denoted name of types.

Reviewed By: shayne-fletcher

Differential Revision: D29925806

fbshipit-source-id: e04af8976e07fcdddb5bdde7202bf1d7743fcd23