Fix fixupVmfpUsages
Summary:
The fixupVmfpUsages logic assumes that a given frame pointer will only
be used in a regchain with one other frame pointer. This is because
once the regchain is inserted into, we never remove it (and will fail
if we try to do a conflicting insert). Once we encounter the popvmfp,
we need to remove the corresponding entry. However, this means we
cannot keep a single regchain map for the entire unit. The regchain
map is now flow sensitive. Instead of keeping one, do the usual thing
and keep one per-block and use dataflow.

The current logic all works now, but it starts to assert once I
implement some optimizations and we start to GVN frame pointers more
aggressively.

Reviewed By: mofarrell

Differential Revision: D30185397

fbshipit-source-id: a4c615173a40a25010c1757f3b90d7fb3fd9e3e6