Tweak HhasRecord C++: Field, Replace 'a str with Str<'arena>
Summary: It will no longer be reasonable to retain references into a parse AST from an HHBC AST when the latter are being served to hhvm. So, it's not enough to just replace the `'a str` in `Field` with an `Str<'a>`, rather we need to install a copy of it into a `Str<'arena>`.

Differential Revision: D30192518

fbshipit-source-id: 355da00cfb5b2e627d0e7808cc1769b6c5c6d549