Remove redundant stores of known values in load-elim
Summary:
Currently, we fail to optimize IR of the following form:

```
Store<Loc1> t1   # PureStore      { Loc1, t1 }
MayLoad<Loc1>    # GeneralEffects { Loc1, AEmpty, AEmpty, AEmpty }
Store<Loc1> t1   # PureStore      { Loc1, t1 }
MayLoad<Loc1>    # GeneralEffects { Loc1, AEmpty, AEmpty, AEmpty }
```

The second PureStore can be safely elided, as we are storing to a location with an available, identical value. Store elimination does not remove this store: it does not track values, and both stores are consumed.

Load elimination tracks availability and known values of locations. Use this information to determine when a PureStore is provably redundant, and eliminate the store in that case.

Reviewed By: ricklavoie

Differential Revision: D30185333

fbshipit-source-id: 2bc8057692e3582addd54646bc4e2fd29ec400d2