Tany/Terr/Tobject logger
Summary:
A pseudo TAST check that is conditionally enabled to log `Tany`, `Terr`, and `Tobject` types. These are bad types that we want to keep track of.

It gathers information both from function/method signatures and expression types. Individual appearances of `Tany`, `Terr`, and `Tobject` are gathered in the `info` data type. Please see the documentation on this data type for the details of the information we collect.

We also log types that contain these types, e.g., a `vec` of `Tany`s.

Reviewed By: vsiles

Differential Revision: D29591054

fbshipit-source-id: 0bc66cba2afcdc3d8b30295e6c6ac0379393ba38