Enable more DCE of dead 86reified_prop accesses in HHBBC
Summary:
There were a couple of things preventing us from dead accesses of the
86reified_prop. The first was that we were treating Dims very
conservatively with regards to props. Improve that a bit mark obvious
safe cases as being nothrow.

The second issue was the special logic to try to force the type of the
86reified_prop. Not only was this incomplete, it's technically a
monotonicity violation since the type can oscillate between InitCell
and the specialized type (it was usually InitCell) during DCE. Instead
of (incorrectly) just asserting the type, make it work correctly with
the existing analysis. Treat 86reified_init as a special init function
(like the others), and do the necessary analysis with
CheckReifiedGenericMismatch. One final thing is dealing with the prop
specially when loosing for serialization.

While I was doing it, I noticed we also failed to DCE certain
sequences involving control flow. The reason was a phase ordering
issued, solved by doing an additonal optimization in DCE-opts.

Reviewed By: mofarrell

Differential Revision: D30204971

fbshipit-source-id: a62bc5ddfbecd6023d9eebb0f588c12a68602a69