Tweak HHBC AST C++: Don't sign hhbc-ast.h, no need
Summary:
As noted in https://www.internalfb.com/diff/D30220880?dst_version_fbid=389757465906837&transaction_fbid=314774243488089, there's no need to sign the generated header since now we have buck handling caching for us.

Since we don't sign anymore, don't inject an `generated <<...>>` seed into the generated header and instead, replace it with a warning that the generated file is not  to be manually modified.

Reviewed By: paulbiss

Differential Revision: D30231081

fbshipit-source-id: 381044a80453959657691278d7f0db7d0b475148