Traverse temporary variable assignments in identify symbol service
Summary:
Previously, we did not traverse the desugared expression nor the temporary variable assignments in the identify symbol service.
Not traversing the desugared expression prevents from populating the unityped runtime code to the user. However, because the temporary variable assignments is used in the typing logic, it can be useful to traverse these statements as part of the service.

Reviewed By: Wilfred

Differential Revision: D29781219

fbshipit-source-id: fc34e9f4ee2acc0b896e430a1c78504c2910978e