Fix access for type constant from unresolved type
Summary:
Based on the late solve behavior in D29447669 (https://github.com/facebook/hhvm/commit/7f8089d813b447987ae21d1e51d78121f44198c2), I think it is possible to get rid of the FIXMEs in Expression Trees today.

The current problem is that, when we desugar expression trees, we get a bunch of function pointers floating around that don't have their type variable bound or constrained to anything. This is despite the fact that looking at the actual code, these type variables should be constrained by the usage of the function. Let's take a look at an example:

```
abstract class MyBox {
  abstract const type TInner;
}

class IntBox extends MyBox {
  const type TInner = ExampleInt;
}

async function setState<T as MyBox, TVal>(
  ExampleContext $_visitor,
): Awaitable<ExprTree<Code, Code::TAst, (function(T, TVal): void)>>
where
  TVal = T::TInner {
  throw new \Exception();
}

function test(ExprTree<Code, Code::TAst, IntBox> $x): void {
  Code`setState(${$x}, 1)`;
}
```

In this example, one expects that the type variables associated with the call to `setState` is lower bounded to `IntBox` and `ExampleInt`. This is true. However, when we desugar this expression tree, we get desugared code that looks like:

```
function test(ExprTree<Code, Code::TAst, IntBox> $x): void {
  (
    () ==> {
      $_splice0 = $x;
      return Code::makeTree(
        darray[
          "path" => __FILE__,
          "start_line" => 24,
          "end_line" => 24,
          "start_column" => 7,
          "end_column" => 25,
        ],
        darray[
          "splices" => dict["\$_splice0" => $_splice0],
          // This is a problematic usage of the function pointer
          "functions" => vec[setState<>],
          "static_methods" => vec[],
        ],
        (
          (Code $_v) ==> {
            return $_v->visitCall(
              darray[
                "path" => __FILE__,
                "start_line" => 24,
                "end_line" => 24,
                "start_column" => 7,
                "end_column" => 25,
              ],
              $_v->visitGlobalFunction(
                darray[
                  "path" => __FILE__,
                  "start_line" => 24,
                  "end_line" => 24,
                  "start_column" => 7,
                  "end_column" => 25,
                ],
                setState<>,
...
```

So while the `setState` call in the virtualized code is being appropriately constrained and bound, the `setState` pointers floating around are not, and it is those places where the typing error is coming from.

This diff performs the following hack. Instead of referring to a new function pointer all over the place, assign that function pointer to a local variable, then use the local variable everywhere instead of the function pointer. This way, due to the late solve behavior in D29447669 (https://github.com/facebook/hhvm/commit/7f8089d813b447987ae21d1e51d78121f44198c2), we wait until we use the function pointer in the virtualized expression, which binds the type variables before solving. This way, we don't get into the circumstance where there are function pointers around that are unbound.

For purposes of visualizing this change, consider the new behavior compared the desugared code above:

```
function test(ExprTree<Code, Code::TAst, IntBox> $x): void {
  (
    () ==> {
      $_splice0 = $x;
      // Currently unbound function pointer
      $_fp0 = setState<>;
      return Code::makeTree(
        darray[
          "path" => __FILE__,
          "start_line" => 24,
          "end_line" => 24,
          "start_column" => 7,
          "end_column" => 25,
        ],
        darray[
          "splices" => dict["\$_splice0" => $_splice0],
          // Refers to the currently unbound function pointer
          "functions" => vec[$_fp0],
          "static_methods" => vec[],
        ],
        (
          (Code $_v) ==> {
            return $_v->visitCall(
              darray[
                "path" => __FILE__,
                "start_line" => 24,
                "end_line" => 24,
                "start_column" => 7,
                "end_column" => 25,
              ],
              $_v->visitGlobalFunction(
                darray[
                  "path" => __FILE__,
                  "start_line" => 24,
                  "end_line" => 24,
                  "start_column" => 7,
                  "end_column" => 25,
                ],
                $_fp0,
...
```

Then, when the virtualized code is checked:

```
Code::symbolType($_fp0)(${$x}, Code::intType())
```

the variable `$_fp0` has constraints added and it is now possible to solve for the type variables. Since we only refer to this function by name once, all places where we use it are appropriately solved and we don't get the error anymore!

(The following implementation is pretty hacky)

Probably not really for shipping at the moment. Just to get feedback on the approach.

Reviewed By: Wilfred

Differential Revision: D29506989

fbshipit-source-id: 45da879c5dbe6d75982952614215492f4ef08cac