Add ReadOnlyOp to BaseL
Summary:
The cases
```
$x[0] =  <rhs> - CheckROCOW

```
could actually be ok when `$x` is readonly if it is also COW. In order to check this, we just need to allow the `CheckROCOW` ReadOnlyOp on BaseL.

Since HackC already knows the readonlyness of `$x`, HHVM will just check COW.

Note: The only ReadOnlyOp allowed on BaseL is `CheckROCOW` because if $x was mutable, we don't need any check. It's only when $x is readonly so choosing `CheckROCOW` over `CheckMutROCOW` just encodes this information.

Reviewed By: paulbiss

Differential Revision: D30072237

fbshipit-source-id: 9f72c4831aa550af73619ef93e4eef59d9273953