Implement basic memory rematerialization in vasm-graph-color
Summary:
Add support to the vasm-graph-color rematerialization logic to deal
with memory loads. For this to work, we need to know if the involved
memory location has since been written to. We do not have anything
like vasm memory effects, but have them on the HHIR level.

Use the HHIR memory effects to determine (conservatively) the memory
locations any given vasm instruction may read or write to. Vasm
instructions have an origin field, linking them to their source HHIR
instruction. Using this, we calculate memory effects and from that the
memory locations touched. This is conservative, since not all HHIR
memory locations map cleanly into the vasm level, and just because the
HHIR instruction has a memory effect, doesn't mean *every* vasm
instruction for that HHIR instruction has that effect. However, it's
"good enough" for most cases. We need to deal specially with vasm
instructions with no origin.

The rematerialization logic is modified as follows: The usual defining
instruction logic is the same, except certain memory load instructions
are now considered "acceptable". These are instructions which have a
"PureLoad" memory effect and only read from a single location. To
determine if the found load can be rematerialized at the given program
point, we walk backwards through the program. For every instruction
encountered along the way, we check its associated write locations. If
it conflicts with the original load, we're done and cannot
rematerialize. Otherwise, if we don't encouter any conflicting writes
on any path back to the original load, it's safe. This logic is very
similar to the defining instruction logic.

NB: This makes it so it's even more important to set a vasm
instruction's origin properly.

Differential Revision: D30185747

fbshipit-source-id: 6fa07ddda15d44bb7d70731b963d59589262b0f3