Remove all tmps spanning call restrictions
Summary:
We currently forbid SSATmps from spanning Call instructions. The
reason is that a Call instruction is going to clobber all of the
physical registers, and therefore anything live across the Call will
get spilled. Instead of spilling, it should just get reloaded from
it's original location.

This is true, but it inhibits other optimizations. For example, if a
SSATmp is kept alive across a Call, load and store elim may be able to
sink any usage of it onto side-exits, and we can remove it entirely
from the main path. Furthermore, the restriction is a bit
inconsistent. We do it for Call instructions, but not any of the HHIR
instructions which can cause spilling. In practice, we spill a lot
anyways.

Now that the register allocator can rematerialize simple mem loads,
one of the biggest pros for this restriction is gone. Remove it
entirely and let all optimization passes freely move SSATmps across
Calls. From examination, it appears the register allocator is
successful in removing a great amount of spills introduced.

Reviewed By: oulgen, jano

Differential Revision: D30185808

fbshipit-source-id: f28ed64c126b1358b0f747f98933016eae594d3f