assert that naming_table is empty at start of full init
Summary:
I'm struggling to understand why the naming-table is sometimes reporting duplicate names on a clean repo when it does a fallback to full init.

This diff is a desperate grasp at a straw. I wonder if maybe the naming table already had something in it prior to adding all names?

This diff involves folding over the entire naming table. That would be slow if it were the full naming table. However, the point of this diff is that the naming table at this point in time should be empty, and so folding should be cheap.

Differential Revision: D30271271

fbshipit-source-id: 5634734995c5fad24cf60323990e496dc4578eae