in case of naming_invariant_violation, log full FileInfo
Summary:
This diff is grasping at a straw for why we sometimes (rarely) get INVARIANT_VIOLATION_BUG "naming_duplicate_same".

My hunch is that maybe the FileInfo.t is corrupt, e.g. includes the exact same entry twice in one of its lists.

With this diff, if we get that "naming_duplicate_same" error, we'll log it to scuba.

Differential Revision: D30271312

fbshipit-source-id: 3542f6c3399b603921dbf7c7d5a03ea9708f6df9