Property access virtualization and desugaring
Summary:
There is a desire to mimic getting and setting state variables in Expression Tree DSLs. One option for enabling this is by enabling property access in Expression Trees.

```
// Setting state
Code`(MyState $state) ==> {
  $state->data = 1;
}`

// Getting state
Code`(MyState $state) ==> {
  do_something($state->data);
}`
```

## Desugaring

Properties would be desugared as

```
Code::makeTree(
  ...
  (Code $v) ==> {
    $v->visitLambda(
      ...
      vec[
        $v->visitPropertyAccess(..., $v->visitLocal(..., '$state'), 'data'),
      ],
    ),
  },
)
```

Where the visitor API would be

```
  public function visitPropertyAccess(
    ?ExprPos $_,
    Code::TAst $recv,
    string $property_name,
  ): Code::TAst {
    throw new Exception();
  }
```

Where the `$recv` argument is the object from which the property is being accessed and the `$property_name` argument is the string representation of the property name.

## Typing/Virtualization

For typing and virtualization, we are not performing any additional virtualization for property access. Instead, we rely on the class typing directly to inform the typing of property access in expression trees.

For instance, one can write the following class type
```
abstract class MyState {
  public int $prop;
}
```

The property itself is not virtualized within the expression tree, so in a world where the `Code` Expression Tree int is virtualized to `CodeInt`, the following would be an error:

```
Code`(MyState $x) ==> {
  $x->prop = 1;
}`;
// Error, trying to assign a CodeInt to a property of type int
```

So one would want to write out a class definition as:

```
abstract class MyState {
  public CodeInt $prop;
}

Code`(MyState $x) ==> {
  // No error
  $x->prop = 1;
}`;
```

## Restrictions

Under this model, we are currently not accepting dynamic property access in the form of

```
Code`(MyState $state) ==> {
  $dyn = 'prop';
  // Banned
  $state->$dyn = 1;
}`
```
We are also not currently accepting nullsafe property access in which the receiver object may be null.
```
Code`(?MyState $state) ==> {
  // Banned
  $state?->prop;
}`
```

While this does allow the possibility of calling properties as though they contained lambdas:

```
Code`(?MyState $state) ==> {
  // Possible
  ($state->callable)();
}`
```

This does not allow direct method calls

```
Code`(?MyState $state) ==> {
  // Banned
  $state->my_method();
}`
```

## Additional notes

This allows for chained property access

```
Code`(MyState $state) ==> {
  $state->nested_state->my_prop;
}`
```

This syntax is enabled for accessing properties on any other allowed expression, including splices:

```
Code`() ==> {
  create_state()->my_prop;
}`

Code`() ==> {
  ${ $spliced_var }->my_prop;
}`
```

Reviewed By: Wilfred

Differential Revision: D29953410

fbshipit-source-id: 5203d4e824fef14d6e14d57c67f97b76fbc3344f