List assignment
Summary:
Basic list assignment checks

Note that something like

```
list($a, $b) = vec[readonly new Foo(), new Foo()];
```
will make $a and $b both readonly, even though technically $b is mutable. This is because the entire vec literal is considered readonly, and we use that readonlyness to assign to the list. I don't think having an exception for literals/pairs etc would be worth doing here, and the extra complexity does not lead to much better expressiveness, so I'm leaving it to be more conservative.

Reviewed By: kavoor

Differential Revision: D30195214

fbshipit-source-id: e5e488dfc8e32d781cb1ef2fa0ed29a976e6100b