Fix deep property accesses in combination with array get
Summary:
This diff has a weird algorithm in it whose correctness I haven't 100% proved, but I'm pretty sure its right, so please check my work.

It concerns code similar to this:

```
$x->prop[0] = new C();
```
Here, if $x->prop is a readonly vec, we need the edit to succeed, since vec is a value type. *But if $x itself is readonly*, the compiler should fail, since we are editing the property of $x.

Thus, when looking at an lval that ends in an array_get (i.e. the bracket syntax):

```
<expr>[] = <rval>
```

<expr> here can really only be one of three values:
- It can be a property access, i.e. `<inner_exp>->prop`
- It can be another array access, i.e. `<inner_exp>[$i]`
- It can be a local variable, i.e. `$x`, which is the base case.
- I don't think it can legally be anything else in Hack, but any other case I haven't thought of is likely the same as the base case.

Thus, we split the code for checking assignment into two categories: assigning to a local, which is handled specifically through editing the context and all that jazz, and assigning to a nonlocal. List expressions and other collections are a TODO for now. For the nonlocal case, we look at property accesses and array accesses.

- Array accesses `assign(<expr>[], <rval>)` can be handled by recursion: the expression <expr>[] = <rval> is valid if and only if `assign(<expr>, <rval>)` for readonly analysis (obviously in normal types, this isn't true, but for readonly, with only two types that propogate, this is correct!).

- Property access `assign(<expr>->prop, <rval>)`  is valid if `<expr>` is mutable.

- Anything else, `assign($x, <rval>)` is a base case, where we check the readonlyness of $x vs. rval.

Examples:

```
// $x is readonly before this line
$x->prop[0] = readonly new C();
```
Checking this involves:
```
assign($x->prop[0], readonly new C())
=> assign($x->prop, readonly new C())
=> error, $x is readonly
```

```
// $x is mutable before this line
$x->prop[0] = readonly new C();
```
Checking this involves:
```
assign($x->prop[0], readonly new C())
=> assign($x->prop, readonly new C())
=> Ok, make rhs explicitly readonly if it isn't already. Make
```

Reviewed By: kavoor

Differential Revision: D30108230

fbshipit-source-id: 9490d199ad1df94dec4ed5c317dbf15a3a0b7bb8