Add dynamic-type of $this dependent coeffects to IMemoizeParam::getInstanceKey
Summary:
We added it and never used it. This doesn't result in Hack errors.

Maybe someone can remind me why we never used it?

Why add it? Mostly just the reserve the name in case someone wants to be polymorphic on getInstanceKey in the futture. The alternative I think is to just remove the ctx const.

Someone asked about it so I checked up on it again: https://fb.workplace.com/groups/hackforhiphop/posts/6457612120954015

Reviewed By: rodmk

Differential Revision: D30309416

fbshipit-source-id: ff794be75e103d081023a36b33a594e178f47933