Move over dependent contexts to T/[...] naming
Summary:
This consolidates all of the generated type parameters for effect-polymorphic functions into a single form prefixed by a non-denotable `T/`. The generic that replaces a parameter `$a`'s type hint will be `T/$a` and the generics that replace context hints will be
- `ctx $f` => `T/[ctx $f]`
- `$a::C` => `T/[$a::C]`

This establishes a consistent prefix that can be used to determine whether a type parameter is explicit in the source or generated by the lowerer without tracking separate lists of tparams, and it provides a friendlier view in case a type parameter leaks to the front-end, since it closely matches the source. Also, this simplifies the reconstruction of contexts for error messages.

Misc: I've done a minor refactor of Decl_hint.hint and callers to separate out hints that are useful only for coeffects (e.g. Hvar) to prevent leaking.

Reviewed By: losvald

Differential Revision: D30097147

fbshipit-source-id: d3ac5d21030c7df3edd8efa8c94c46062d096980