Allow for phis with rvmfp() as a source in vasm-graph-color
Summary:
Previously we didn't allow any phis between ignored registers
and non-ignored registers since they didn't occur. We have some diffs
now that can make these occur (specifically involving rvmfp(), so
loosen the requirements a bit. We now allow a phi to take an ignored
register as a source, even if the dest is non-ignored. We deal with
this condition by simplying copying the source into a temporary and
using that as the phi source.

Reviewed By: colavitam

Differential Revision: D30334567

fbshipit-source-id: 46c8be358beb4bf00fd55a2e635b9d90c9ff8181