Add flushing in Config_file printing
Summary:
The output in Config_file_common.apply_overrides is getting printed out of order because Rust `println` doesn't flush automatically and isn't using the same buffer as OCaml's Printf (or something along those lines). This makes the output very misleading.

Add flushes so that this output appears in the correct order.

Reviewed By: bobrenjc93

Differential Revision: D30321010

fbshipit-source-id: c6bf695a70dbfb0d9562e09784f41f3182d5fb50