Really basic readonly property emitter
Summary:
Emit readonly immediates for most basic property access case. Add TODOs for others, which I'll add later as time goes on

Note: This diff will require that the runtime have a flag to control whether to turn on runtime support for properties, and uses that flag. HackC will emit the correct immediates for readonly expressions with or without readonly enforcement being on.

Reviewed By: kavoor

Differential Revision: D30115664

fbshipit-source-id: 7dca938354fc1a6aab8fcef19aecf0a7124216e9