Refactor the extraction process
Summary:
Previously, as part of the desugaring process for Expression Trees, there were three separate passes. The first extracted expressions from splices, the second extracted function pointers and static method pointers. And the final pass rewrote the Expression Tree.

This diff combines all three passes over the Expression Tree, performing the rewrite (the virtualization and the desugaring) as well as extracting the spliced extractions and the function pointers and replacing them with temporary variables.

Reviewed By: Wilfred

Differential Revision: D29810170

fbshipit-source-id: 3b365f7676da1a3be3815f1a07bfa8bd7c405896