Annotate contexts for ReflectionProperty::getValue/setValue
Summary:
These APIs work for both instance properties as well as static properties, which means that they need to be annotated with the union of the capabilities needed for each.

For getting values, reading an instance prop is pure, but reading static props needs `read_globals`, so we end up with `[read_globals]`.

For setting values, writing to an instance prop requires `write_props`, and writing to a static prop requires `globals`, so we end up with `[globals, write_props]`.

There's a case to be made for specific getter/setters that are specific to static and instance properties, to get the narrowest set of capabilities required, but we can expose those later as needed.

Reviewed By: viratyosin

Differential Revision: D30203725

fbshipit-source-id: 9b3d284d6e6b69df299c95644cc40f3e2cc24f47