Improved bidirectional checking for ImmVector literals and others
Summary: Bidirectional type-checking (the "expected type") is currently unaware of direct ancestors of `ImmVector` and friends for checking literal expressions such as `ImmVector { ... }`. This means, for example, that when returning a literal `ImmVector` as a `ConstVector`, we don't make use of the expected type when checking the elements of the `ImmVector`. We lose a little bit of checking (see new test cases), but perhaps more importantly we lose performance, and have to construct a union unnecessarily, for example, in `ImmVector { 'a', 2, true }` returned as `ConstVector<mixed>`.

Reviewed By: enetsee

Differential Revision: D30391266

fbshipit-source-id: cef90e47bb34e0380511919723f327e183a55a84