Support for CheckROCOW
Summary:
This diff adds support for CheckROCOW Readonly immediate.
 ---

Recently we added `CheckROCOW` as a ReadOnly immediate to support this case:

```
$a->p[0][0][0[] = readonly $x;
```
Where HHVM must enforce that p is readonly and COW.
This is different from `CheckMutROCOW`, which targets this case:
```
$a->p[0][0][0[] = $x;
```
Where HHVM must enforce that p is mutable *unless* it is readonly and COW.

The only difference in terms of *where* `CheckROCOW` can exist compared to `CheckMutROCOW` is on the BaseL bytecode instruction. Only `CheckROCOW` can be found there (HackC knows readonlyness of locals, and if the value is mutable, we don't need any checks and therefore would never need to emit `CheckMutROCOW`).

Reviewed By: oulgen

Differential Revision: D30315483

fbshipit-source-id: 8922091f5a01be7f7efe65e1838ef6390f6c85b6