Experiment: Fix all AST types to `#[repr(C)]`
Summary:
RFC:

This is a necessary (but not sufficient) part of being able to pass our Rust AST values (e.g. decls) to C++. It has been verified to have no observable effects on the tests or performance.

The change is to add `repr(C)` to the `attrs_and_vis` calculation in the `type_declaration` function in `convert_type_decl.ml`.

The effect is that all generated Rust types will now have their representations fixed to `repr(C)`. This allows for interoperability of these types with the C programming language standardizing alignment and layout in a compatible fashion. See https://doc.rust-lang.org/reference/type-layout.html#the-c-representation for details.

Any concerns?

Differential Revision: D30342116

fbshipit-source-id: 049d521f73dd9a9f6a05c8d0a221fb7ef441c173