Readonly base property enforcement
Summary:
This changes emit_base to emit correct readonlyop immediates for certain property cases.

When emitting a base, if the base is a ArrayGet, we should pass the existing readonlyop requirement down.  (For example, in an lval position, we pass in the readonlyop CheckROCOW or CheckMutROCOW depending on the rhs's readonlyness.

 If the base is a ObjGet, we must enforce that the readonlyop is Mutable for all of the inner object.

This diff does *not* implement the BaseL case (i.e. checking `$x[] = readonly new Foo()` that $x is copy on write), as that requires us to fill in the readonly expression on the left side during inference. We will do that in a later diff.

Reviewed By: kavoor

Differential Revision: D30348020

fbshipit-source-id: f79ff64643b63cc004de867b518ed5b1b7e4f59e