Use required-ancestors as well as normal ancestors in type refinement
Summary: When refining a variable of static type `A<T>` with a test against `B<_>`, we propagate the type parameter e.g. `B<T>` extends `A<T>` directly. Unfortunately we do not currently do this if the relationship is `require extends` rather than a direct `extends`. To fix this, we use `Cls.requires_ancestor` as well as `Cls.has_ancestor`.

Differential Revision: D30394856

fbshipit-source-id: b98f719fa8c8318e091ade32eaddc1b09e483c46