Add a new opaque resource type to hhvm
Summary:
This new resource type is currently only exposed via a `__SystemLib` interface. The plan is to use it as an opaque type in places where we may want to eventually specialize the type of a value in a new hack feature and in the interim don't want to expose any implementation details about it.

It's currently planned for use by the new enum atom feature.

Reviewed By: ricklavoie

Differential Revision: D30349379

fbshipit-source-id: 8f92035c0a27b80aa33ab391abf7e07f3ffa7d62