Implement projection of context constants off reified generics
Summary:
This implements projection of context constants from reified generics in an identical way to the projection of contexts from parameters, namely

```
// given
abstract class class A {
  abstract const ctx C;
}

// argument
function f(A $a)[$a::C]: void {}

==>

function f<
  T/$a,
  T/[$a::C]
>(
  T/$a $a
): void where T/[$a::C] = T/$a :: C {}

// reified generic
function f<reify T as A>()[T::C] {}

==>

function f<
  reify T as A,
  T/[T::C]
>(): void where T/[T::C] = T :: C {}
```
As with parameter-dependent contexts, these contexts are restricted to a single projected constant.

*Misc:* I've expanded the tracking of generics in the lowerer to functions and erased generics. This enables detection of projections from reified generics as they have the same structure as normal type projections:

`Haccess (Happly n, ["C"])`

where n is a type name. Currently we rely on Naming to convert `Happly` to `Habstr` when the name is a type parameter in scope. The direct decl parser does a better job by directly emitting type parameter hints as `Tgeneric` decls (as opposed to `Tapply`). I think with a medium-sized refactor (out of scope for this diff) we can use the new tracking to switch to emitting `Habstr` when lowering hints and simplify a bunch of downstream logic.

Reviewed By: jamesjwu

Differential Revision: D30097146

fbshipit-source-id: 25aa61d33ad3c50fd9d403c5b2059b29bfe65900