Must be Enclosed in Readonly Expression Exception
Summary:
Per D30119495 (https://github.com/facebook/hhvm/commit/70aaa08773dc919e5c349e895a7ad4f7160132a1)
It would be nice if
```
<?hh //strict

<<file: __EnableUnstableFeatures('readonly')>>

class P {
  public function __construct(public readonly ?P $prop=null) { }
}

<<__EntryPoint>>
function main() {
  $p = new P();
  $p->prop = new P();
  $z = readonly $p->prop; // should be okay
  $w = $p->prop; // not okay
}
```
could output
```
"This property is readonly, but isn't enclosed in a readonly expression"
```
when `$p` "Must Be Mutable" as the developer likely meant to wrap the rhs in a readonly expression, not that they meant for the property to be mutable.

The rule here is that the rhs must be mutable unless its wrapped in a readonly expression, in which case we emit no checks.

Reviewed By: oulgen

Differential Revision: D30322399

fbshipit-source-id: 1db0329f511f69aaeb62dc80b9fe65a3114c0b57